Make use of new proton-j APIs to improve client performance
Use the APIs available in new proton-j releases to improve overall
performance on send.  Allows buffers to be sized ahead of large writes
to avoid multiple resizes and exposes the outgoing buffer as a
non-readonly byte buffer to allow for direct array copy to netty.

This is related to issue #87 update to proton-j 0.30.0