ReadableBuffer string reads must advance the read index
The contract of the ReadableBuffer from proton-j indicates that
the buffer implementations of the read string methods should
advance the read index. The current implementation is not doing
so but currently doesn't cause any issue because proton-j does
its string decoding with a slice, future versions might not do
this and the result would be failed decodes due to the index
not being advanced.